[C-934] Migrate user-list sagas to common
