+++ packages/web/src/common/store/user-list/following/sagas.ts
@@ -2,17 +2,17 @@ import {
   ID,
   User,
   cacheUsersSelectors,
-  followersUserListSelectors,
   UserListSagaFactory,
-  followersUserListActions,
-  FOLLOWERS_USER_LIST_TAG as USER_LIST_TAG
+  followingUserListActions,
+  followingUserListSelectors,
+  FOLLOWING_USER_LIST_TAG
 } from '@audius/common'
 import { put, select } from 'typed-redux-saga'
 
-import { watchFollowersError } from 'common/store/user-list/followers/errorSagas'
-import { createUserListProvider } from 'components/user-list/utils'
-const { getFollowersError } = followersUserListActions
-const { getId, getUserList, getUserIds } = followersUserListSelectors
+import { watchFollowingError } from 'common/store/user-list/following/errorSagas'
+import { createUserListProvider } from 'common/store/user-list/utils'
+const { getId, getUserList, getUserIds } = followingUserListSelectors
+const { getFollowingError } = followingUserListActions
 const { getUser } = cacheUsersSelectors
 
 const provider = createUserListProvider<User>({
@@ -25,7 +25,7 @@ const provider = createUserListProvider<User>({
     currentUserId,
     apiClient
   }) => {
-    const users = await apiClient.getFollowers({
+    const users = await apiClient.getFollowing({
       currentUserId,
       profileUserId: entityId,
       limit,
@@ -35,30 +35,31 @@ const provider = createUserListProvider<User>({
   },
   selectCurrentUserIDsInList: getUserIds,
   canFetchMoreUsers: (user: User, combinedUserIDs: ID[]) =>
-    combinedUserIDs.length < user.follower_count,
-  includeCurrentUser: (u) => u.does_current_user_follow
+    combinedUserIDs.length < user.followee_count,
+
+  includeCurrentUser: (_) => false
 })
 
 function* errorDispatcher(error: Error) {
   const id = yield* select(getId)
   if (id) {
-    yield* put(getFollowersError(id, error.message))
+    yield* put(getFollowingError(id, error.message))
   }
 }
 
-function* getFollowers(currentPage: number, pageSize: number) {
+function* getFollowing(currentPage: number, pageSize: number) {
   const id: number | null = yield* select(getId)
   if (!id) return { userIds: [], hasMore: false }
   return yield* provider({ id, currentPage, pageSize })
 }
 
 const userListSagas = UserListSagaFactory.createSagas({
-  tag: USER_LIST_TAG,
-  fetchUsers: getFollowers,
+  tag: FOLLOWING_USER_LIST_TAG,
+  fetchUsers: getFollowing,
   stateSelector: getUserList,
   errorDispatcher
 })
 
 export default function sagas() {
-  return [userListSagas, watchFollowersError]
+  return [userListSagas, watchFollowingError]
 }
