+++ packages/web/src/store/sagas.ts
@@ -37,7 +37,14 @@ import smartCollectionPageSagas from 'common/store/smart-collection/sagas'
 import socialSagas from 'common/store/social/sagas'
 import artistRecommendationsSagas from 'common/store/ui/artist-recommendations/sagas'
 import reactionSagas from 'common/store/ui/reactions/sagas'
+import favoritePageSagas from 'common/store/user-list/favorites/sagas'
+import followersPageSagas from 'common/store/user-list/followers/sagas'
+import followingPageSagas from 'common/store/user-list/following/sagas'
+import mutualsPageSagas from 'common/store/user-list/mutuals/sagas'
 import notificationUsersPageSagas from 'common/store/user-list/notifications/sagas'
+import repostPageSagas from 'common/store/user-list/reposts/sagas'
+import supportingPageSagas from 'common/store/user-list/supporting/sagas'
+import topSupportersPageSagas from 'common/store/user-list/top-supporters/sagas'
 import addToPlaylistSagas from 'components/add-to-playlist/store/sagas'
 import changePasswordSagas from 'components/change-password/store/sagas'
 import firstUploadModalSagas from 'components/first-upload-modal/store/sagas'
@@ -51,17 +58,10 @@ import rewardsPageSagas from 'pages/audio-rewards-page/store/sagas'
 import collectionSagas from 'pages/collection-page/store/sagas'
 import deactivateAccountSagas from 'pages/deactivate-account-page/store/sagas'
 import deletedSagas from 'pages/deleted-page/store/sagas'
-import favoritePageSagas from 'pages/favorites-page/sagas'
-import followersPageSagas from 'pages/followers-page/sagas'
-import followingPageSagas from 'pages/following-page/sagas'
 import historySagas from 'pages/history-page/store/sagas'
-import mutualsPageSagas from 'pages/mutuals-page/sagas'
 import remixesSagas from 'pages/remixes-page/store/sagas'
-import repostPageSagas from 'pages/reposts-page/sagas'
 import searchPageSagas from 'pages/search-page/store/sagas'
 import settingsSagas from 'pages/settings-page/store/sagas'
-import supportingPageSagas from 'pages/supporting-page/sagas'
-import topSupportersPageSagas from 'pages/top-supporters-page/sagas'
 import uploadSagas from 'pages/upload-page/store/sagas'
 import { initInterface } from 'services/native-mobile-interface/helpers'
 import webAnalyticsSagas from 'store/analytics/sagas'
